refactor: move consts from test to mock
